Refactor ListItemType and update transaction item models
Replaces the old ListItemType implementation with a simplified version supporting only text, custom, and empty cases. Updates transaction-related view models and scene logic to use the new ListItemType and TransactionItemModel cases. Removes the ListItemViewable protocol and related code, and adapts AddressListItemView to use ListItemImageView directly. This refactor streamlines list item configuration and improves code clarity.
